Mod. 2 training styles moved to TrainingStyles.js Co-authored-by: Austin Huisinga <huisingaaustin@gmail.com> Co-authored-by: Lisa Cook <lisabpink@gmail.com> Co-authored-by: Raj Upadhyaya <xrajupadhyayax@gmail.com>
