+++ src/components/Training/module-2/components/module2-2.js
@@ -7,12 +7,14 @@ import {
   Media,
   Video,
   Form,
-  Label,
+  Module2Label,
   Input,
   Submit,
-
+  Section, 
+  List, 
+  Module2Step 
 } from '../../TrainStyles'
-import { Section, List, Step } from '../styles'
+
 import { SimpleStep } from './index'
 import Progress from "../../TrainingModuleContainer";
 
@@ -68,7 +70,7 @@ export const Module22 = ({ handleBack, errors, touched }) => {
             <div key={index}>
               <SimpleStep title={step.title} text={step.text} />
               {index === 0 && (
-                <Step>
+                <Module2Step>
                   <h3>Start with simple, small tasks.</h3>
                   <MediaWrapper>
                     <div>
@@ -98,7 +100,7 @@ export const Module22 = ({ handleBack, errors, touched }) => {
                     budget or time demands. You can always add tasks, but it's
                     best not to over commit.
                   </p>
-                </Step>
+                </Module2Step>
               )}
             </div>
           ))}
@@ -106,10 +108,10 @@ export const Module22 = ({ handleBack, errors, touched }) => {
       </Section>
       <Title>Understanding of responsibility: Complete before proceeding</Title>
       <Form>
-        <Label>
+        <Module2Label>
           Please summarize your responsibility (relationally and practically) in
           this match with a sentence or two. *
-        </Label>
+        </Module2Label>
         <Input
           component='textarea'
           id='m2_q2'
